musicmeta/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musicmeta"
version = "1.0.0"
description = "跨平台音乐元数据编辑器 - Cross-platform Music Metadata Editor"
authors = [
    {name = "MusicMeta Team"}
]
readme = "README.md"
requires-python = ">=3.8"
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: MacOS X",
    "Environment :: Win32 (MS Windows)",
    "Environment :: X11 Applications",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "mutagen==1.47.0",
    "Pillow==10.1.0",
    "customtkinter==5.2.1",
    "darkdetect==0.8.0",
    "packaging==23.2",
    "requests==2.31.0",
    "beautifulsoup4==4.12.2",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "flake8>=6.0",
]
windows = [
    "pywin32==306",
]
android = [
    "android-storage==1.0.0",
]

[project.urls]
Homepage = "https://github.com/musicmeta/musicmeta"
Documentation = "https://github.com/musicmeta/musicmeta/wiki"
Repository = "https://github.com/musicmeta/musicmeta.git"

[project.scripts]
musicmeta = "src.main:main"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
musicmeta = ["assets/*", "assets/**/*"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_functions = "test_*"
